backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
msgpack==1.0.8
pillow==10.4.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
PyJWT==2.8.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
redis==5.0.8
requests==2.31.0
setuptools==73.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.1
wrapt==1.16.0
zope.interface==7.0.1